Move messages to their own dialogs.ts module
